src/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.33.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic-settings==2.8.1
aiofiles==24.1.0
langchain==0.2.17
langchain-community==0.2.19
PyMuPDF==1.24.11
